refactor blog theme components organization
