+++ packages/docusaurus-plugin-content-blog/src/options.ts
@@ -34,8 +34,8 @@ export const DEFAULT_OPTIONS: PluginOptions = {
   showReadingTime: true,
   blogTagsPostsComponent: '@theme/BlogTagsPostsPage',
   blogTagsListComponent: '@theme/BlogTagsListPage',
-  blogAuthorsPostsComponent: '@theme/BlogAuthorsPostsPage',
-  blogAuthorsListComponent: '@theme/BlogAuthorsListPage',
+  blogAuthorsPostsComponent: '@theme/Blog/Pages/BlogAuthorsPostsPage',
+  blogAuthorsListComponent: '@theme/Blog/Pages/BlogAuthorsListPage',
   blogPostComponent: '@theme/BlogPostPage',
   blogListComponent: '@theme/BlogListPage',
   blogArchiveComponent: '@theme/BlogArchivePage',
